scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chonkie[all]>=1.5.1",
    "ebooklib>=0.20",
    "google-generativeai>=0.8.6",
    "pymupdf>=1.26.7",
]
